+++ gems.rb
@@ -7,6 +7,9 @@ source "https://rubygems.org"
 
 gemspec
 
+# Use the protocol fixes while their release is pending:
+gem "protocol-grpc", git: "https://github.com/socketry/protocol-grpc.git", ref: "cb474e449da5864a91206a680c2543e84619ea89"
+
 group :maintenance, optional: true do
 	gem "bake-gem"
 	gem "bake-modernize"